+++ lib/src/model/locale.dart
@@ -1,0 +1,553 @@
+// ignore_for_file: constant_identifier_names
+
+/// Enumerated locale
+enum Locale {
+  /// Abkhazian
+  AB,
+
+  /// Afar
+  AA,
+
+  /// Afrikaans
+  AF,
+
+  /// Akan
+  AK,
+
+  /// Albanian
+  SQ,
+
+  /// Amharic
+  AM,
+
+  /// Arabic
+  AR,
+
+  /// Aragonese
+  AN,
+
+  /// Armenian
+  HY,
+
+  /// Assamese
+  AS,
+
+  /// Avaric
+  AV,
+
+  /// Avestan
+  AE,
+
+  /// Aymara
+  AY,
+
+  /// Azerbaijani
+  AZ,
+
+  /// Bambara
+  BM,
+
+  /// Bashkir
+  BA,
+
+  /// Basque
+  EU,
+
+  /// Belarusian
+  BE,
+
+  /// Bengali
+  BN,
+
+  /// Bislama
+  BI,
+
+  /// Bosnian
+  BS,
+
+  /// Breton
+  BR,
+
+  /// Bulgarian
+  BG,
+
+  /// Burmese
+  MY,
+
+  /// Catalan, Valencian
+  CA,
+
+  /// Chamorro
+  CH,
+
+  /// Chechen
+  CE,
+
+  /// Chichewa, Chewa, Nyanja
+  NY,
+
+  /// Chinese
+  ZH,
+
+  /// Church Slavic, Old Slavonic, Church Slavonic, Old Bulgarian, Old Church Slavonic
+  CU,
+
+  /// Chuvash
+  CV,
+
+  /// Cornish
+  KW,
+
+  /// Corsican
+  CO,
+
+  /// Cree
+  CR,
+
+  /// Croatian
+  HR,
+
+  /// Czech
+  CS,
+
+  /// Danish
+  DA,
+
+  /// Divehi, Dhivehi, Maldivian
+  DV,
+
+  /// Dutch, Flemish
+  NL,
+
+  /// Dzongkha
+  DZ,
+
+  /// English
+  EN,
+
+  /// Esperanto
+  EO,
+
+  /// Estonian
+  ET,
+
+  /// Ewe
+  EE,
+
+  /// Faroese
+  FO,
+
+  /// Fijian
+  FJ,
+
+  /// Finnish
+  FI,
+
+  /// French
+  FR,
+
+  /// Western Frisian
+  FY,
+
+  /// Fulah
+  FF,
+
+  /// Gaelic, Scottish Gaelic
+  GD,
+
+  /// Galician
+  GL,
+
+  /// Ganda
+  LG,
+
+  /// Georgian
+  KA,
+
+  /// German
+  DE,
+
+  /// Greek, Modern (1453–)
+  EL,
+
+  /// Kalaallisut, Greenlandic
+  KL,
+
+  /// Guarani
+  GN,
+
+  /// Gujarati
+  GU,
+
+  /// Haitian, Haitian Creole
+  HT,
+
+  /// Hausa
+  HA,
+
+  /// Hebrew
+  HE,
+
+  /// Herero
+  HZ,
+
+  /// Hindi
+  HI,
+
+  /// Hiri Motu
+  HO,
+
+  /// Hungarian
+  HU,
+
+  /// Icelandic
+  IS,
+
+  /// Ido
+  IO,
+
+  /// Igbo
+  IG,
+
+  /// Indonesian
+  ID,
+
+  /// Interlingua (International Auxiliary Language Association)
+  IA,
+
+  /// Interlingue, Occidental
+  IE,
+
+  /// Inuktitut
+  IU,
+
+  /// Inupiaq
+  IK,
+
+  /// Irish
+  GA,
+
+  /// Italian
+  IT,
+
+  /// Japanese
+  JA,
+
+  /// Javanese
+  JV,
+
+  /// Kannada
+  KN,
+
+  /// Kanuri
+  KR,
+
+  /// Kashmiri
+  KS,
+
+  /// Kazakh
+  KK,
+
+  /// Central Khmer
+  KM,
+
+  /// Kikuyu, Gikuyu
+  KI,
+
+  /// Kinyarwanda
+  RW,
+
+  /// Kirghiz, Kyrgyz
+  KY,
+
+  /// Komi
+  KV,
+
+  /// Kongo
+  KG,
+
+  /// Korean
+  KO,
+
+  /// Kuanyama, Kwanyama
+  KJ,
+
+  /// Kurdish
+  KU,
+
+  /// Lao
+  LO,
+
+  /// Latin
+  LA,
+
+  /// Latvian
+  LV,
+
+  /// Limburgan, Limburger, Limburgish
+  LI,
+
+  /// Lingala
+  LN,
+
+  /// Lithuanian
+  LT,
+
+  /// Luba-Katanga
+  LU,
+
+  /// Luxembourgish, Letzeburgesch
+  LB,
+
+  /// Macedonian
+  MK,
+
+  /// Malagasy
+  MG,
+
+  /// Malay
+  MS,
+
+  /// Malayalam
+  ML,
+
+  /// Maltese
+  MT,
+
+  /// Manx
+  GV,
+
+  /// Maori
+  MI,
+
+  /// Marathi
+  MR,
+
+  /// Marshallese
+  MH,
+
+  /// Mongolian
+  MN,
+
+  /// Nauru
+  NA,
+
+  /// Navajo, Navaho
+  NV,
+
+  /// North Ndebele
+  ND,
+
+  /// South Ndebele
+  NR,
+
+  /// Ndonga
+  NG,
+
+  /// Nepali
+  NE,
+
+  /// Norwegian
+  NO,
+
+  /// Norwegian Bokmål
+  NB,
+
+  /// Norwegian Nynorsk
+  NN,
+
+  /// Sichuan Yi, Nuosu
+  II,
+
+  /// Occitan
+  OC,
+
+  /// Ojibwa
+  OJ,
+
+  /// Oriya
+  OR,
+
+  /// Oromo
+  OM,
+
+  /// Ossetian, Ossetic
+  OS,
+
+  /// Pali
+  PI,
+
+  /// Pashto, Pushto
+  PS,
+
+  /// Persian
+  FA,
+
+  /// Polish
+  PL,
+
+  /// Portuguese
+  PT,
+
+  /// Punjabi, Panjabi
+  PA,
+
+  /// Quechua
+  QU,
+
+  /// Romanian, Moldavian, Moldovan
+  RO,
+
+  /// Romansh
+  RM,
+
+  /// Rundi
+  RN,
+
+  /// Russian
+  RU,
+
+  /// Northern Sami
+  SE,
+
+  /// Samoan
+  SM,
+
+  /// Sango
+  SG,
+
+  /// Sanskrit
+  SA,
+
+  /// Sardinian
+  SC,
+
+  /// Serbian
+  SR,
+
+  /// Shona
+  SN,
+
+  /// Sindhi
+  SD,
+
+  /// Sinhala, Sinhalese
+  SI,
+
+  /// Slovak
+  SK,
+
+  /// Slovenian
+  SL,
+
+  /// Somali
+  SO,
+
+  /// Southern Sotho
+  ST,
+
+  /// Spanish, Castilian
+  ES,
+
+  /// Sundanese
+  SU,
+
+  /// Swahili
+  SW,
+
+  /// Swati
+  SS,
+
+  /// Swedish
+  SV,
+
+  /// Tagalog
+  TL,
+
+  /// Tahitian
+  TY,
+
+  /// Tajik
+  TG,
+
+  /// Tamil
+  TA,
+
+  /// Tatar
+  TT,
+
+  /// Telugu
+  TE,
+
+  /// Thai
+  TH,
+
+  /// Tibetan
+  BO,
+
+  /// Tigrinya
+  TI,
+
+  /// Tonga (Tonga Islands)
+  TO,
+
+  /// Tsonga
+  TS,
+
+  /// Tswana
+  TN,
+
+  /// Turkish
+  TR,
+
+  /// Turkmen
+  TK,
+
+  /// Twi
+  TW,
+
+  /// Uighur, Uyghur
+  UG,
+
+  /// Ukrainian
+  UK,
+
+  /// Urdu
+  UR,
+
+  /// Uzbek
+  UZ,
+
+  /// Venda
+  VE,
+
+  /// Vietnamese
+  VI,
+
+  /// Volapük
+  VO,
+
+  /// Walloon
+  WA,
+
+  /// Welsh
+  CY,
+
+  /// Wolof
+  WO,
+
+  /// Xhosa
+  XH,
+
+  /// Yiddish
+  YI,
+
+  /// Yoruba
+  YO,
+
+  /// Zhuang, Chuang
+  ZA,
+
+  /// Zulu
+  ZU,
+}