+++ test/nominatim_geocoding_test.dart
@@ -7,17 +7,17 @@ void main() async {
   test('Check forward geocoding', () async {
     final result = await NominatimGeocoding.to.forwardGeoCoding(
       const Address(
-        city: 'Braunschweig',
+        city: 'Brunswick',
         postalCode: 38120,
       ),
     );
-    expect(result.address.city, 'Braunschweig');
+    expect(result.address.city, 'Brunswick');
   });
   test('Check reverse geocoding', () async {
     final result = await NominatimGeocoding.to.reverseGeoCoding(
       const Coordinate(
           latitude: 52.27429313260939, longitude: 10.523078303155874),
     );
-    expect(result.address.city, 'Braunschweig');
+    expect(result.address.city, 'Brunswick');
   });
 }